.history/requirements_20240113115907.txt
tensorflow-text>=2.11
einops